+++ config/application.rb
@@ -22,5 +22,27 @@ module SimRep
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
     config.web_console.development_only = false # srrrrrryyyyyy
+
+    # We want to set up a custom logger which logs to STDOUT.
+    # Docker expects your application to log to STDOUT/STDERR and to be ran
+    # in the foreground.
+    config.log_level = :debug
+    config.log_tags  = [:subdomain, :uuid]
+
+    if ENV["RAILS_LOG_TO_STDOUT"].present?
+      logger           = ActiveSupport::Logger.new(STDOUT)
+      logger.formatter = config.log_formatter
+      config.logger = ActiveSupport::TaggedLogging.new(logger)
+    end
+
+    # Since we're using Redis for Sidekiq, we might as well use Redis to back
+    # our cache store. This keeps our application stateless as well.
+    config.cache_store = :redis_store, ENV['CACHE_URL']
+
+    # If you've never dealt with background workers before, this is the Rails
+    # way to use them through Active Job. We just need to tell it to use Sidekiq.
+    config.active_job.queue_adapter = :sidekiq
+
+    # ...
   end
 end